amunicja, zmiana inv frames chargera

+++ ProjektGra/AnimationClip.cpp
@@ -133,17 +133,18 @@ InvFramesCharger::InvFramesCharger() :
 
 InvFramesCharger::~InvFramesCharger()
 {
+	
 }
 
 void InvFramesCharger::function()
 {
 	if (!fully_charged)
 	{
-		if (current_charge < 100)
+		if (current_charge < 100.f)
 		{
 			current_charge += charge_value;
 
-			float f = static_cast<float>(current_charge) / 100.f;
+			float f = current_charge / 100.f;
 
 			if (f > 1.f)
 			{
@@ -162,7 +163,7 @@ void InvFramesCharger::function()
 
 void InvFramesCharger::set_ready()
 {
-	current_charge = 0;
+	current_charge = 0.f;
 	fully_charged = true;
 	DisplayHP::set_percentage(1.f, DisplayHP::dodge_text);
 }
